Async communication

+++ Services/Catalog/Catalog.Application/Responses/ProductForCreationDto.cs
@@ -10,5 +10,5 @@ public record ProductForCreationDto
     public string? ImageFile { get; init; }
     public ProductBrand? Brands { get; init; }
     public ProductType? Types { get; init; }
-    public decimal? price { get; init; }
+    public int? price { get; init; }
 }
